requirements.txt
tk
Pillow
matplotlib
numpy
